+++ src/config.rs
@@ -70,6 +70,8 @@ pub enum ModeConfiguration {
         port: u16,
         #[serde(default)]
         prefix: Option<String>,
+        #[serde(default)]
+        http_timeout: Option<u32>,
     },
 }