+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ==================
+
+- Added configuration for upload HTTP request timeout `http_timeout` (#26)
 - (Breaking) Remove `ptr` option from configuration, replaced with `prefix` which can be set to
   a desired API URL prefix - `prefix = "ptr"` replaces the functionality of `ptr = true`
 - (Breaking) The default `port` is now 443 and `ssl` now defaults to true (previous default was